permit summary
